+++ OutputterCtemplate.cpp
@@ -36,7 +36,7 @@ void OutputterCtemplate::print(cppdb::result &res, std::string &outstr) {
 	ctemplate::TemplateDictionary dict(template_name);
 	while(res.next()) {
 		ctemplate::TemplateDictionary* single_entry = dict.AddSectionDictionary("ELEMENT");
-		for(int i=0;i<OutputterBasic::parameters.size();i++) {
+		for(size_t i=0; i<OutputterBasic::parameters.size(); i++) {
 			std::string tmp;
 			res >> tmp;
 			single_entry->SetValue(parameters.at(i),converters.at(i) ? converters.at(i)->getValue(tmp) : tmp);
